replace G2Snapshot with DBSnapshot
