+++ G2Snapshot.py
@@ -8,17 +8,25 @@ import random
 import signal
 import sys
 import time
+import math
 from datetime import datetime, timedelta
 import configparser
+
+#--concurrency
+from multiprocessing import Process, Queue, Value
+from queue import Empty, Full
+import threading
 
 #--senzing python classes
 try: 
     import G2Paths
+    from G2Database import G2Database
     from G2Product import G2Product
     from G2Engine import G2Engine
     from G2IniParams import G2IniParams
     from G2ConfigMgr import G2ConfigMgr
     from G2Exception import G2Exception
+    from G2Diagnostic import G2Diagnostic
 except:
     print('')
     print('Please export PYTHONPATH=<path to senzing python directory>')
@@ -26,44 +34,394 @@ except:
     sys.exit(1)
 
 #---------------------------------------
-def nextExportRecord(exportHandle, exportHeaders = None):
-    rowString = bytearray()
-    rowData = g2Engine.fetchNext(exportHandle, rowString)
-    if not(rowData):
+def queue_read(queue): 
+    try: return queue.get(True, 1)
+    except Empty:
+        time.sleep(.01)
         return None
 
-    if exportHeaders == 'JSON':
-        return json.loads(rowData)
+#---------------------------------------
+def queue_write(queue, message): 
+    while True:
+        try: queue.put(message, True, 1)
+        except Full:
+            time.sleep(.01)
+            continue
+        break
 
-    try: rowData = next(csv.reader([rowString.decode()[0:-1]]))
-    except: 
-        print(' err: ' + rowString.decode())        
-        return None
+#---------------------------------------
+def setup_entity_queue_db(thread_id, threadStop, entity_queue, resume_queue):
+    local_dbo = G2Database(g2dbUri)
+    if not local_dbo.success:
+        print('\nCould not connect to database\n')
+        with shutDown.get_lock():
+            shutDown.value = 1
+        return
+    process_entity_queue_db(thread_id, threadStop, entity_queue, resume_queue, local_dbo)
+    local_dbo.close()
 
-    if exportHeaders: 
-        return dict(zip(exportHeaders, rowData))
+#---------------------------------------
+def setup_entity_queue_api(thread_count, threadStop, entity_queue, resume_queue):
+    #--multi-threaded API
+    try: 
+        g2Engine = G2Engine()
+        g2Engine.initV2('G2Snapshot', iniParams, False)
+    except G2Exception as ex:
+        printWithNewLines(f'G2Exception: {ex}', 'B')
+        with shutDown.get_lock():
+            shutDown.value = 1
+        return
+    thread_list = []
+    for thread_id in range(thread_count):
+        thread_list.append(threading.Thread(target=process_entity_queue_api, args=(thread_id, threadStop, entity_queue, resume_queue, g2Engine)))
+    for thread in thread_list:
+        thread.start()
+    for thread in thread_list:
+        thread.join()
+    g2Engine.destroy()
+
+#---------------------------------------
+def setup_entity_queue_api_single(thread_id, threadStop, entity_queue, resume_queue):
+    #--single-threaded API
+    try: 
+        g2Engine = G2Engine()
+        g2Engine.initV2('G2Snapshot%s' % thread_id, iniParams, False)
+    except G2Exception as ex:
+        printWithNewLines(f'G2Exception: {ex}', 'B')
+        with shutDown.get_lock():
+            shutDown.value = 1
+        return
+    process_entity_queue_api(thread_id, threadStop, entity_queue, resume_queue, g2Engine)
+    g2Engine.destroy()
+
+#---------------------------------------
+def process_entity_queue_db(thread_id, threadStop, entity_queue, resume_queue, local_dbo):
+    while threadStop.value == 0: #or entity_queue.empty() == False:
+        queue_data = queue_read(entity_queue)
+        if queue_data:
+            #print('read entity_queue %s' % row)  
+            resume_rows = get_resume_db(local_dbo, queue_data)
+            if resume_rows:
+                queue_write(resume_queue, resume_rows)
+    #print('process_entity_queue %s shut down with %s left in the queue' % (thread_id, entity_queue.qsize()))
+
+#---------------------------------------
+def process_entity_queue_api(thread_id, threadStop, entity_queue, resume_queue, g2_engine):
+    while threadStop.value == 0: #entity_queue.empty() == False:
+        queue_data = queue_read(entity_queue)
+        if queue_data:
+            #print('read entity_queue %s' % row)  
+            resume_rows = get_resume_api(g2_engine, queue_data)
+            if resume_rows:
+                queue_write(resume_queue, resume_rows)
+    #print('process_entity_queue2 %s shut down with %s left in the queue' % (thread_id, entity_queue.qsize()))
+
+#---------------------------------------
+def setup_resume_queue(statPack, thread_id, threadStop, resume_queue):
+
+    if exportCsv:
+        columnHeaders = []
+        columnHeaders.append('RESOLVED_ENTITY_ID')
+        columnHeaders.append('RELATED_ENTITY_ID')
+        columnHeaders.append('MATCH_LEVEL')
+        columnHeaders.append('MATCH_KEY')
+        columnHeaders.append('DATA_SOURCE')
+        columnHeaders.append('RECORD_ID')
+        try: 
+            exportFileHandle = open(csvFilePath, 'a')
+            exportFileHandle.write(','.join(columnHeaders) + '\n')        
+        except IOError as err: 
+            print('\nERROR: cannot write to %s \n%s\n' % (csvFilePath, err))
+            with shutDown.get_lock():
+                shutDown.value = 1
+            return
+    else:
+        exportFileHandle = None
+            
+    process_resume_queue(thread_id, threadStop, resume_queue, statPack, exportFileHandle)
+
+    if exportCsv:
+        exportFileHandle.close()
+
+
+#---------------------------------------
+def process_resume_queue(thread_id, threadStop, resume_queue, statPack, exportFileHandle):
+    while threadStop.value == 0: #or resume_queue.empty() == False:
+        queue_data = queue_read(resume_queue)
+        if queue_data:
+            #print('read resume_queue', row)
+
+            #--its a resume to process
+            if type(queue_data) == list: 
+                statPack = process_resume(statPack, queue_data, exportFileHandle)
+
+            #--its a status write request
+            else:
+                statPack = write_stat_pack(statPack, queue_data)
+
+    #print('process_resume_queue %s shut down with %s left in the queue' % (thread_id, resume_queue.qsize()))
+
+#---------------------------------------
+def get_resume_db(local_dbo, resolved_id):
+
+    resume_rows = []
+
+    sqlEntities = 'select ' + \
+                  ' a.RES_ENT_ID as RESOLVED_ENTITY_ID, '\
+                  ' a.ERRULE_ID, '\
+                  ' a.MATCH_KEY, '\
+                  ' b.DSRC_ID, '\
+                  ' c.RECORD_ID '\
+                  'from RES_ENT_OKEY a '\
+                  'join OBS_ENT b on b.OBS_ENT_ID = a.OBS_ENT_ID '\
+                  'join DSRC_RECORD c on c.ENT_SRC_KEY = b.ENT_SRC_KEY and c.DSRC_ID = b.DSRC_ID and c.ETYPE_ID = b.ETYPE_ID '\
+                  'where a.RES_ENT_ID = ?'
+
+    if not exportCsv: #--don't need related record_id
+        sqlRelations = 'select '\
+                       ' a.RES_ENT_ID as RESOLVED_ENTITY_ID, '\
+                       ' a.REL_ENT_ID as RELATED_ENTITY_ID, '\
+                       ' b.LAST_ERRULE_ID as ERRULE_ID, '\
+                       ' b.IS_DISCLOSED, '\
+                       ' b.IS_AMBIGUOUS, '\
+                       ' b.MATCH_KEY, '\
+                       ' d.DSRC_ID '\
+                       'from RES_REL_EKEY a '\
+                       'join RES_RELATE b on b.RES_REL_ID = a.RES_REL_ID '\
+                       'join RES_ENT_OKEY c on c.RES_ENT_ID = a.REL_ENT_ID '\
+                       'join OBS_ENT d on d.OBS_ENT_ID = c.OBS_ENT_ID '\
+                       'where a.RES_ENT_ID = ?'
+    else:
+       sqlRelations = 'select '\
+                       ' a.RES_ENT_ID as RESOLVED_ENTITY_ID, '\
+                       ' a.REL_ENT_ID as RELATED_ENTITY_ID, '\
+                       ' b.LAST_ERRULE_ID as ERRULE_ID, '\
+                       ' b.IS_DISCLOSED, '\
+                       ' b.IS_AMBIGUOUS, '\
+                       ' b.MATCH_KEY, '\
+                       ' d.DSRC_ID, '\
+                       ' e.RECORD_ID '\
+                       'from RES_REL_EKEY a '\
+                       'join RES_RELATE b on b.RES_REL_ID = a.RES_REL_ID '\
+                       'join RES_ENT_OKEY c on c.RES_ENT_ID = a.REL_ENT_ID '\
+                       'join OBS_ENT d on d.OBS_ENT_ID = c.OBS_ENT_ID '\
+                       'join DSRC_RECORD e on e.ENT_SRC_KEY = d.ENT_SRC_KEY and e.DSRC_ID = d.DSRC_ID and e.ETYPE_ID = d.ETYPE_ID '\
+                       'where a.RES_ENT_ID = ?'
+    
+    queryStartTime = time.time()
+    cursor1 = local_dbo.sqlExec(sqlEntities, [resolved_id,])
+    for rowData in local_dbo.fetchAllDicts(cursor1):
+        rowData = complete_resume_db(rowData)
+        resume_rows.append(rowData)
+    #print('   fetching entities took %s seconds' % str(round(time.time() - queryStartTime,2)))
+
+    #--get relationships if not an orphaned entity_id
+    if resume_rows and relationshipFilter in (2,3):
+        queryStartTime = time.time()
+        cursor1 = local_dbo.sqlExec(sqlRelations, [resolved_id,])
+        for rowData in local_dbo.fetchAllDicts(cursor1):
+            rowData = complete_resume_db(rowData)
+            resume_rows.append(rowData)
+        #print('   fetching relationships took %s seconds' % str(round(time.time() - queryStartTime,2)))
+
+    return resume_rows
+
+#---------------------------------------
+def complete_resume_db(rowData):
+
+    if 'RELATED_ENTITY_ID' not in rowData:
+        rowData['RELATED_ENTITY_ID'] = 0 
+        rowData['IS_DISCLOSED'] = 0 
+        rowData['IS_AMBIGUOUS'] = 0
+    if 'RECORD_ID' not in rowData: 
+        rowData['RECORD_ID'] = 'n/a' 
+
+    try: rowData['DATA_SOURCE'] = dsrcLookup[rowData['DSRC_ID']]['DSRC_CODE']
+    except: rowData['DATA_SOURCE'] = 'unk'
+    try: rowData['ERRULE_CODE'] = erruleLookup[rowData['ERRULE_ID']]['ERRULE_CODE']
+    except: rowData['ERRULE_CODE'] = 'unk'
+
+    if rowData['RELATED_ENTITY_ID'] == 0:
+        rowData['MATCH_LEVEL'] = 1 if rowData['ERRULE_CODE'] else 0
+        rowData['MATCH_CATEOGRY'] = 'RESOLVED'
+    elif rowData['IS_DISCLOSED'] != 0:
+        rowData['MATCH_LEVEL'] = 11
+        rowData['MATCH_CATEGORY'] = 'DISCLOSED_RELATION'
+    elif rowData['IS_AMBIGUOUS'] != 0:
+        rowData['MATCH_LEVEL'] = 11
+        rowData['MATCH_CATEGORY'] = 'AMBIGUOUS_MATCH'
+    else:
+        try: rowData['MATCH_LEVEL'] = erruleLookup[rowData['ERRULE_ID']]['RTYPE_ID']
+        except: rowData['MATCH_LEVEL'] = 3
+        if rowData['MATCH_LEVEL'] == 2:
+            rowData['MATCH_CATEGORY'] = 'POSSIBLE_MATCH'
+        else:          
+            rowData['MATCH_CATEGORY'] = 'POSSIBLY_RELATED'
 
     return rowData
 
 #---------------------------------------
-def initDataSourceStats(dataSource1, dataSource2 = None):
-    if not dataSource2:
-        statPack['DATA_SOURCES'][dataSource1] = {}
-        statPack['DATA_SOURCES'][dataSource1]['ENTITY_COUNT'] = 0
-        statPack['DATA_SOURCES'][dataSource1]['RECORD_COUNT'] = 0
-        statPack['DATA_SOURCES'][dataSource1]['SINGLE_COUNT'] = 0
-        statPack['DATA_SOURCES'][dataSource1]['SINGLE_SAMPLE'] = []
-        for statType in ['DUPLICATE', 'AMBIGUOUS_MATCH', 'POSSIBLE_MATCH', 'POSSIBLY_RELATED', 'DISCLOSED_RELATION']:
-            statPack['DATA_SOURCES'][dataSource1][statType + '_ENTITY_COUNT'] = 0
-            statPack['DATA_SOURCES'][dataSource1][statType + '_RECORD_COUNT'] = 0
-            statPack['DATA_SOURCES'][dataSource1][statType + '_SAMPLE'] = []
-        statPack['DATA_SOURCES'][dataSource1]['CROSS_MATCHES'] = {}
-    else:
-        statPack['DATA_SOURCES'][dataSource1]['CROSS_MATCHES'][dataSource2] = {}
-        for statType in ['MATCH', 'AMBIGUOUS_MATCH', 'POSSIBLE_MATCH', 'POSSIBLY_RELATED', 'DISCLOSED_RELATION']:
-            statPack['DATA_SOURCES'][dataSource1]['CROSS_MATCHES'][dataSource2][statType + '_ENTITY_COUNT'] = 0
-            statPack['DATA_SOURCES'][dataSource1]['CROSS_MATCHES'][dataSource2][statType + '_RECORD_COUNT'] = 0
-            statPack['DATA_SOURCES'][dataSource1]['CROSS_MATCHES'][dataSource2][statType + '_SAMPLE'] = []
+def get_resume_api(g2Engine, resolved_id):
+
+    resume_rows = []
+
+    getFlags = 0
+    #getFlags = g2Engine.G2_ENTITY_DEFAULT_FLAGS
+    #getFlags = getFlags | g2Engine.G2_ENTITY_INCLUDE_ENTITY_NAME
+    getFlags = getFlags | g2Engine.G2_ENTITY_INCLUDE_RECORD_DATA
+    #getFlags = getFlags | g2Engine.G2_ENTITY_INCLUDE_RECORD_MATCHING_INFO
+    #getFlags = getFlags | g2Engine.G2_ENTITY_INCLUDE_RECORD_FORMATTED_DATA
+    if relationshipFilter in (2,3):
+        getFlags = getFlags | g2Engine.G2_ENTITY_INCLUDE_ALL_RELATIONS
+        #getFlags = getFlags | g2Engine.G2_ENTITY_INCLUDE_RELATED_ENTITY_NAME
+        getFlags = getFlags | g2Engine.G2_ENTITY_INCLUDE_RELATED_MATCHING_INFO
+        getFlags = getFlags | g2Engine.G2_ENTITY_INCLUDE_RELATED_RECORD_SUMMARY
+
+    try: 
+        response = bytearray()
+        retcode = g2Engine.getEntityByEntityIDV2(int(resolved_id), getFlags, response)
+        response = response.decode() if response else ''
+    except G2Exception as err:
+        print(str(err))
+        with shutDown.get_lock():
+            shutDown.value = 1
+        return resume_rows
+
+    if not response:
+        return resume_rows
+
+    try: jsonData = json.loads(response)
+    except:
+        print('json parse error:\n\t ', response)
+        return resume_rows
+
+    for record in jsonData['RESOLVED_ENTITY']['RECORDS']:
+        resume_rows.append({
+            'RESOLVED_ENTITY_ID': jsonData['RESOLVED_ENTITY']['ENTITY_ID'],
+            'RELATED_ENTITY_ID': 0,
+            'MATCH_LEVEL': 1,
+            'IS_AMBIGUOUS': 0,
+            'IS_DISCLOSED': 0,
+            'MATCH_KEY': 'n/a',
+            'DATA_SOURCE': record['DATA_SOURCE'],
+            'RECORD_ID': record['RECORD_ID']
+        })
+    if 'RELATED_ENTITIES' in jsonData:
+        for relation in jsonData['RELATED_ENTITIES']:
+            for record in relation['RECORD_SUMMARY']:
+                resume_rows.append({
+                    'RESOLVED_ENTITY_ID': jsonData['RESOLVED_ENTITY']['ENTITY_ID'],
+                    'RELATED_ENTITY_ID': relation['ENTITY_ID'],
+                    'MATCH_LEVEL': relation['MATCH_LEVEL'],
+                    'IS_AMBIGUOUS': relation['IS_AMBIGUOUS'],
+                    'IS_DISCLOSED': relation['IS_DISCLOSED'],
+                    'MATCH_KEY': relation['MATCH_KEY'],
+                    'DATA_SOURCE': record['DATA_SOURCE'],
+                    'RECORD_ID': 'n/a'
+                })
+
+    return resume_rows
+
+#---------------------------------------
+def process_resume(statPack, resume_rows, exportFileHandle):
+
+    if not resume_rows:
+        return statPack
+
+    categoryTotalStat = {}
+    categoryTotalStat['AMBIGUOUS_MATCH'] = 'TOTAL_AMBIGUOUS_MATCHES'
+    categoryTotalStat['POSSIBLE_MATCH'] = 'TOTAL_POSSIBLE_MATCHES'
+    categoryTotalStat['POSSIBLY_RELATED'] = 'TOTAL_POSSIBLY_RELATEDS'
+    categoryTotalStat['DISCLOSED_RELATION'] = 'TOTAL_DISCLOSED_RELATIONS'
+
+    calcStartTime = time.time()
+
+    entitySize = 0
+    recordList = []
+    resumeData = {}
+
+    #--summarize entity resume 
+    entityID = resume_rows[0]['RESOLVED_ENTITY_ID']
+    for rowData in resume_rows:
+        relatedID = str(rowData['RELATED_ENTITY_ID'])
+        dataSource = rowData['DATA_SOURCE']
+        recordID = rowData['RECORD_ID']
+
+        if relatedID == '0':
+            matchCategory = 'RESOLVED'
+            entitySize += 1
+            recordList.append(dataSource + ':' + recordID)
+        elif rowData['IS_DISCLOSED'] != 0:
+            matchCategory = 'DISCLOSED_RELATION'
+        elif rowData['IS_AMBIGUOUS'] != 0:
+            matchCategory = 'AMBIGUOUS_MATCH'
+        elif rowData['MATCH_LEVEL'] == 2:
+            matchCategory = 'POSSIBLE_MATCH'
+        else:          
+            matchCategory = 'POSSIBLY_RELATED'
+
+        if relatedID not in resumeData:
+            resumeData[relatedID] = {}
+            resumeData[relatedID]['matchCategory'] = matchCategory
+            resumeData[relatedID]['dataSources'] = {}
+
+        if dataSource not in resumeData[relatedID]['dataSources']:
+            resumeData[relatedID]['dataSources'][dataSource] = 1
+        else:
+            resumeData[relatedID]['dataSources'][dataSource] += 1
+
+        if exportFileHandle: 
+             writeCsvRecord(rowData, exportFileHandle)
+
+    #--update entity size breakdown
+    randomIndex = randomSampleIndex()
+    strEntitySize = str(entitySize)
+    if strEntitySize not in statPack['ENTITY_SIZE_BREAKDOWN']:
+        statPack['ENTITY_SIZE_BREAKDOWN'][strEntitySize] = {}
+        statPack['ENTITY_SIZE_BREAKDOWN'][strEntitySize]['COUNT'] = 0
+        statPack['ENTITY_SIZE_BREAKDOWN'][strEntitySize]['SAMPLE'] = []
+    statPack['ENTITY_SIZE_BREAKDOWN'][strEntitySize]['COUNT'] += 1
+    if len(statPack['ENTITY_SIZE_BREAKDOWN'][strEntitySize]['SAMPLE']) < sampleSize:
+        statPack['ENTITY_SIZE_BREAKDOWN'][strEntitySize]['SAMPLE'].append(entityID)
+        statPack['ENTITY_SIZE_SAMPLE_COUNT'] += 1
+    elif randomIndex != 0:
+        statPack['ENTITY_SIZE_BREAKDOWN'][strEntitySize]['SAMPLE'][randomIndex] = entityID
+
+    #--resolved entity stats
+    statPack['TOTAL_ENTITY_COUNT'] += 1
+    statPack['TOTAL_RECORD_COUNT'] += entitySize
+    for dataSource1 in resumeData['0']['dataSources']:
+        recordCount = resumeData['0']['dataSources'][dataSource1]
+
+        #--this just updates entity and record count for the data source
+        statPack = updateStatpack(statPack, dataSource1, None, None, 1, recordCount, None) 
+
+        if recordCount == 1:
+            statPack = updateStatpack(statPack, dataSource1, None, 'SINGLE', 1, 0, entityID)
+        else:
+            statPack = updateStatpack(statPack, dataSource1, None, 'DUPLICATE', 1, recordCount, entityID)
+
+        #--cross matches
+        for dataSource2 in resumeData['0']['dataSources']:
+            if dataSource2 == dataSource1:
+                continue
+            statPack = updateStatpack(statPack, dataSource1, dataSource2, 'MATCH', 1, recordCount, entityID)
+
+        #--related entity stats
+        for relatedID in resumeData:
+            if relatedID == '0':
+                continue
+            matchCategory = resumeData[relatedID]['matchCategory']
+            statPack[categoryTotalStat[matchCategory]] += 1
+            for dataSource2 in resumeData[relatedID]['dataSources']:
+                recordCount = resumeData[relatedID]['dataSources'][dataSource2]
+                if dataSource2 == dataSource1:
+                    dataSource2 = None
+
+                #--avoid double counting within data source (can't be avoided across data sources)
+                if entityID < int(relatedID) or dataSource2:
+                    statPack = updateStatpack(statPack, dataSource1, dataSource2, matchCategory, 1, recordCount, str(entityID) + ' ' + relatedID)
+    return statPack
 
 #---------------------------------------
 def randomSampleIndex():
@@ -73,19 +431,22 @@ def randomSampleIndex():
     return targetIndex
 
 #---------------------------------------
-def updateStatpack(dataSource1, dataSource2, statPrefix, entityCount, recordCount, sampleValue):
+def updateStatpack(statPack, dataSource1, dataSource2, statPrefix, entityCount, recordCount, sampleValue):
     randomIndex = randomSampleIndex()
 
+    if datasourceFilter and dataSource1 != datasourceFilter:
+        return statPack
+
     if dataSource1 not in statPack['DATA_SOURCES']:
-        initDataSourceStats(dataSource1)
+        statPack = initDataSourceStats(statPack, dataSource1)
     if dataSource2 and dataSource2 not in statPack['DATA_SOURCES'][dataSource1]['CROSS_MATCHES']:
-        initDataSourceStats(dataSource1, dataSource2)
+        statPack = initDataSourceStats(statPack, dataSource1, dataSource2)
 
     #--special case for entity/record count at the data source level with no sample value
     if not statPrefix:
         statPack['DATA_SOURCES'][dataSource1]['ENTITY_COUNT'] += entityCount
         statPack['DATA_SOURCES'][dataSource1]['RECORD_COUNT'] += recordCount
-        return
+        return statPack
 
     #--within data source
     if not dataSource2:
@@ -115,9 +476,32 @@ def updateStatpack(dataSource1, dataSource2, statPrefix, entityCount, recordCoun
         elif randomIndex != 0:
             statPack['DATA_SOURCES'][dataSource1]['CROSS_MATCHES'][dataSource2][statPrefix + '_SAMPLE'][randomIndex] = sampleValue
 
+    return statPack
+
 #---------------------------------------
-def writeCsvRecord(csvData):
-    global shutDown
+def initDataSourceStats(statPack, dataSource1, dataSource2 = None):
+
+    if not dataSource2:
+        statPack['DATA_SOURCES'][dataSource1] = {}
+        statPack['DATA_SOURCES'][dataSource1]['ENTITY_COUNT'] = 0
+        statPack['DATA_SOURCES'][dataSource1]['RECORD_COUNT'] = 0
+        statPack['DATA_SOURCES'][dataSource1]['SINGLE_COUNT'] = 0
+        statPack['DATA_SOURCES'][dataSource1]['SINGLE_SAMPLE'] = []
+        for statType in ['DUPLICATE', 'AMBIGUOUS_MATCH', 'POSSIBLE_MATCH', 'POSSIBLY_RELATED', 'DISCLOSED_RELATION']:
+            statPack['DATA_SOURCES'][dataSource1][statType + '_ENTITY_COUNT'] = 0
+            statPack['DATA_SOURCES'][dataSource1][statType + '_RECORD_COUNT'] = 0
+            statPack['DATA_SOURCES'][dataSource1][statType + '_SAMPLE'] = []
+        statPack['DATA_SOURCES'][dataSource1]['CROSS_MATCHES'] = {}
+    else:
+        statPack['DATA_SOURCES'][dataSource1]['CROSS_MATCHES'][dataSource2] = {}
+        for statType in ['MATCH', 'AMBIGUOUS_MATCH', 'POSSIBLE_MATCH', 'POSSIBLY_RELATED', 'DISCLOSED_RELATION']:
+            statPack['DATA_SOURCES'][dataSource1]['CROSS_MATCHES'][dataSource2][statType + '_ENTITY_COUNT'] = 0
+            statPack['DATA_SOURCES'][dataSource1]['CROSS_MATCHES'][dataSource2][statType + '_RECORD_COUNT'] = 0
+            statPack['DATA_SOURCES'][dataSource1]['CROSS_MATCHES'][dataSource2][statType + '_SAMPLE'] = []
+    return statPack
+
+#---------------------------------------
+def writeCsvRecord(csvData, exportFileHandle):
     columnValues = []
     columnValues.append(str(csvData['RESOLVED_ENTITY_ID']))
     columnValues.append(str(csvData['RELATED_ENTITY_ID']))
@@ -125,395 +509,278 @@ def writeCsvRecord(csvData):
     columnValues.append(csvData['MATCH_KEY'][1:] if csvData['MATCH_KEY'] else '')
     columnValues.append(csvData['DATA_SOURCE'])
     columnValues.append(csvData['RECORD_ID'])
-    try: exportFileHandle.write(','.join(columnValues) + '\n')        
+    try: exportFileHandle.write(','.join(columnValues) + '\n')  
     except IOError as err: 
-        print('')
-        print('ERROR: cannot write to %s \n%s' % (exportFilePath, err))
-        print('')
-        shutDown = True
+        print('\nERROR: cannot write to %s \n%s\n' % (csvFilePath, err))
+        with shutDown.get_lock():
+            shutDown.value = 1
 
 #---------------------------------------
 def processEntities():
-    global shutDown
 
-    statPack['TOTAL_RECORD_COUNT'] = 0
-    statPack['TOTAL_ENTITY_COUNT'] = 0
-    statPack['TOTAL_AMBIGUOUS_MATCHES'] = 0
-    statPack['TOTAL_POSSIBLE_MATCHES'] = 0
-    statPack['TOTAL_POSSIBLY_RELATEDS'] = 0
-    statPack['TOTAL_DISCLOSED_RELATIONS'] = 0
-    statPack['DATA_SOURCES'] = {}
-    statPack['ENTITY_SIZE_BREAKDOWN'] = {}
-    entitySizeReviewCount = 0
-
-    categoryTotalStat = {}
-    categoryTotalStat['AMBIGUOUS_MATCH'] = 'TOTAL_AMBIGUOUS_MATCHES'
-    categoryTotalStat['POSSIBLE_MATCH'] = 'TOTAL_POSSIBLE_MATCHES'
-    categoryTotalStat['POSSIBLY_RELATED'] = 'TOTAL_POSSIBLY_RELATEDS'
-    categoryTotalStat['DISCLOSED_RELATION'] = 'TOTAL_DISCLOSED_RELATIONS'
-
-    #--determine the output flags
-    exportFlags = 0
-    exportFlags = g2Engine.G2_EXPORT_INCLUDE_ALL_ENTITIES
-    if relationshipFilter == 1:
-        pass #--don't include any relationships
-    elif relationshipFilter == 2:
-        exportFlags = exportFlags | g2Engine.G2_ENTITY_INCLUDE_POSSIBLY_SAME_RELATIONS
-    else:
-        exportFlags = exportFlags | g2Engine.G2_ENTITY_INCLUDE_ALL_RELATIONS
-
-    if exportType == 'JSON':  #--specify what sections for json
-        print('\nExport type is JSON')
-        exportFlags = exportFlags | g2Engine.G2_ENTITY_INCLUDE_RECORD_DATA
-        exportFlags = exportFlags | g2Engine.G2_ENTITY_INCLUDE_RECORD_MATCHING_INFO 
-        exportFlags = exportFlags | g2Engine.G2_ENTITY_INCLUDE_RELATED_MATCHING_INFO
-        exportFlags = exportFlags | g2Engine.G2_ENTITY_INCLUDE_RELATED_RECORD_SUMMARY
-
-    else:  #--specify what fields for csv
-        exportFields = []
-        exportFields.append('RESOLVED_ENTITY_ID')
-        exportFields.append('RELATED_ENTITY_ID')
-        exportFields.append('MATCH_LEVEL')
-        exportFields.append('MATCH_KEY')
-        exportFields.append('IS_DISCLOSED')
-        exportFields.append('IS_AMBIGUOUS')
-        exportFields.append('ERRULE_CODE')
-        exportFields.append('DATA_SOURCE')
-        exportFields.append('RECORD_ID')
-
-    #--initialize the export
-    print('\nQuerying entities ...')
-    try: 
-        if exportType == 'JSON':
-            exportHandle = g2Engine.exportJSONEntityReport(exportFlags)
+    newStatPack = True
+    if os.path.exists(statsFilePath):
+        statPack = json.load(open(statsFilePath))
+        if 'PROCESS' in statPack:
+            priorStatus = statPack['PROCESS']['STATUS']
+            lastEntityID = statPack['PROCESS']['LAST_ENTITY_ID'] if type(statPack['PROCESS']['LAST_ENTITY_ID']) == int else 0
         else:
-            exportHandle = g2Engine.exportCSVEntityReportV2(",".join(exportFields), exportFlags)
-            exportHeaders = nextExportRecord(exportHandle)
+            priorStatus = 'Unknown'
+            lastEntityID = 0
+        print ('\n%s snapshot file exists with %s entities processed!' % (priorStatus, statPack['TOTAL_ENTITY_COUNT']))
 
-    except G2Exception as err:
-        print('\n%s\n' % str(err))
+        if priorStatus != 'Complete' and lastEntityID != 0:
+            reply = input('\nDo you want to pick up where it left off (yes/no)? ')
+            if reply in ['y','Y', 'yes', 'YES']:
+                newStatPack = False
+
+        if newStatPack:
+            reply = input('\nAre you sure you want to overwrite it (yes/no)? ')
+            if reply not in ['y','Y', 'yes', 'YES']:
+                with shutDown.get_lock():
+                    shutDown.value = 1
+                return
+        print()
+
+    if newStatPack and os.path.exists(csvFilePath):
+        print('\nThe %s file still exists.  Please either rename or remove it as well.\n' % csvFilePath)
+        with shutDown.get_lock():
+            shutDown.value = 1
+        return
+
+    if newStatPack:
+        statPack = {}
+        statPack['SOURCE'] = 'G2Snapshot'
+        statPack['PROCESS'] = {}
+        statPack['PROCESS']['STATUS'] = 'Incomplete'
+        statPack['PROCESS']['START_TIME'] = datetime.now().strftime('%m/%d/%Y %H:%M:%S')
+        statPack['PROCESS']['LAST_ENTITY_ID'] = 0
+        statPack['TOTAL_RECORD_COUNT'] = 0
+        statPack['TOTAL_ENTITY_COUNT'] = 0
+        statPack['TOTAL_AMBIGUOUS_MATCHES'] = 0
+        statPack['TOTAL_POSSIBLE_MATCHES'] = 0
+        statPack['TOTAL_POSSIBLY_RELATEDS'] = 0
+        statPack['TOTAL_DISCLOSED_RELATIONS'] = 0
+        statPack['DATA_SOURCES'] = {}
+        statPack['ENTITY_SIZE_BREAKDOWN'] = {}
+        statPack['ENTITY_SIZE_SAMPLE_COUNT'] = 0
+
+    g2Dbo = G2Database(g2dbUri)
+    if not g2Dbo.success:
+        print('\nCould not connect to database\n')
         return 1
+
+    entity_queue = Queue(threadCount * 10)
+    resume_queue = Queue(threadCount * 10)
+    thread_list = []
+
+    print('starting %s %s threads ...' % (threadCount, ('api' if use_api else 'database')))
+    process_list = []
+
+    if not use_api: #--sql only mode
+        for thread_id in range(threadCount - 1):
+            process_list.append(Process(target=setup_entity_queue_db, args=(thread_id, threadStop, entity_queue, resume_queue)))
+
+    else: #--api
+        unusedThreads = threadCount - 1
+        while unusedThreads > 0:
+            if unusedThreads > 16:
+                thisThreadCount = 16
+                unusedThreads -= 16
+            else:
+                thisThreadCount = unusedThreads
+                unusedThreads = 0
+            process_list.append(Process(target=setup_entity_queue_api, args=(thisThreadCount, threadStop, entity_queue, resume_queue)))
+
+    process_list.append(Process(target=setup_resume_queue, args=(statPack, 99, threadStop, resume_queue)))
+    for process in process_list:
+        process.start()
+
+    procStartTime = time.time()
+
+    if not datasourceFilter:
+        maxEntityId = g2Dbo.fetchRow(g2Dbo.sqlExec('select max(RES_ENT_ID) from RES_ENT'))[0]
+        sql0 = 'select RES_ENT_ID from RES_ENT where RES_ENT_ID between ? and ?'
+    else:
+        print(f'\ndetermining entity_id range for {datasourceFilter} ... ', end='', flush=True)
+        sql = 'select  ' \
+              ' min(b.RES_ENT_ID), ' \
+              ' max(b.RES_ENT_ID) ' \
+              'from OBS_ENT a ' \
+              'join RES_ENT_OKEY b on b.OBS_ENT_ID = a.OBS_ENT_ID ' \
+              'where a.DSRC_ID = ' + str(datasourceFilterID)
+        minEntityId, maxEntityId = g2Dbo.fetchRow(g2Dbo.sqlExec(sql))
+        print(f'from {minEntityId} to {maxEntityId}\n')
+        if newStatPack: 
+            statPack['PROCESS']['LAST_ENTITY_ID'] = minEntityId - 1
+
+        sql0 = 'select distinct' \
+               ' a.RES_ENT_ID ' \
+               'from RES_ENT_OKEY a ' \
+               'join OBS_ENT b on b.OBS_ENT_ID = a.OBS_ENT_ID ' \
+               'where a.RES_ENT_ID between ? and ? and b.DSRC_ID = ' + str(datasourceFilterID)
 
     batchStartTime = time.time()
-    batchEntitySizeSum = 0
-    batchRelationCountSum = 0
-    maxEntitySize = 0
-    maxRelationCount = 0
-    if exportType == 'JSON':
-        exportRecord = nextExportRecord(exportHandle, exportType)
-    else:
-        exportRecord = nextExportRecord(exportHandle, exportHeaders)
-    while exportRecord:
+    entityCount = 0
+    batchEntityCount = 0
 
-        entitySize = 0
-        recordList = []
-        resumeData = {}
+    begEntityId = statPack['PROCESS']['LAST_ENTITY_ID'] + 1
+    endEntityId = begEntityId + chunkSize
 
-        #--gather the records in a common structure
-        rowList = []
-        if exportType == 'JSON':
-            resolvedEntityID = str(exportRecord['RESOLVED_ENTITY']['ENTITY_ID'])
-            for dsrcRecord in exportRecord['RESOLVED_ENTITY']['RECORDS']:
-                rowData = {}
-                rowData['RESOLVED_ENTITY_ID'] = resolvedEntityID
-                rowData['RELATED_ENTITY_ID'] = '0'
-                rowData['MATCH_LEVEL'] = str(dsrcRecord['MATCH_LEVEL'])
-                rowData['MATCH_KEY'] = dsrcRecord['MATCH_KEY']
-                rowData['IS_DISCLOSED'] = '0'
-                rowData['IS_AMBIGUOUS'] = '0'
-                rowData['DATA_SOURCE'] = dsrcRecord['DATA_SOURCE']
-                rowData['RECORD_ID'] = dsrcRecord['RECORD_ID']
-                rowList.append(rowData)
-            for relatedEntity in exportRecord['RELATED_ENTITIES'] if 'RELATED_ENTITIES' in exportRecord else []:
-                for dsrcSummary in relatedEntity['RECORD_SUMMARY']:
-                    rowData = {}
-                    rowData['RESOLVED_ENTITY_ID'] = resolvedEntityID
-                    rowData['RELATED_ENTITY_ID'] = str(relatedEntity['ENTITY_ID'])
-                    rowData['MATCH_LEVEL'] = str(relatedEntity['MATCH_LEVEL'])
-                    rowData['MATCH_KEY'] = relatedEntity['MATCH_KEY']
-                    rowData['IS_DISCLOSED'] = str(relatedEntity['IS_DISCLOSED'])
-                    rowData['IS_AMBIGUOUS'] = str(relatedEntity['IS_AMBIGUOUS'])
-                    rowData['DATA_SOURCE'] = dsrcSummary['DATA_SOURCE']
-                    rowData['RECORD_ID'] = None
-                    rowList.append(rowData)
-            exportRecord = nextExportRecord(exportHandle, exportType)
-        else: #--csv
-            entityID = exportRecord['RESOLVED_ENTITY_ID']
-            while exportRecord and exportRecord['RESOLVED_ENTITY_ID'] == entityID:
-                rowList.append(exportRecord)
-                exportRecord = nextExportRecord(exportHandle, exportHeaders)
+    while True:
+        if not datasourceFilter:
+            print('Getting entities from %s to %s ...' % (begEntityId, endEntityId))
+        else:
+            print('Getting entities from %s to %s with %s records ...' % (begEntityId, endEntityId, datasourceFilter))
+        entity_rows = g2Dbo.fetchAllRows(g2Dbo.sqlExec(sql0, (begEntityId, endEntityId)))
 
-        #--summarize entity resume 
-        entityID = rowList[0]['RESOLVED_ENTITY_ID']
-        for rowData in rowList:
-            relatedID = rowData['RELATED_ENTITY_ID']
-            dataSource = rowData['DATA_SOURCE']
-            recordID = rowData['RECORD_ID']
+        if entity_rows:
+            last_row_entity_id = entity_rows[len(entity_rows)-1][0]
+        for entity_row in entity_rows:
+            queue_write(entity_queue, entity_row[0])
+            #print('put queue1 %s' % row['RES_ENT_ID'])
 
-            if rowData['RELATED_ENTITY_ID'] == '0':
-                matchCategory = 'RESOLVED'
-                entitySize += 1
-                recordList.append(dataSource + ':' + recordID)
-            elif rowData['IS_DISCLOSED'] != '0':
-                matchCategory = 'DISCLOSED_RELATION'
-            elif rowData['IS_AMBIGUOUS'] != '0':
-                matchCategory = 'AMBIGUOUS_MATCH'
-            elif rowData['MATCH_LEVEL'] == '2':
-                matchCategory = 'POSSIBLE_MATCH'
-            else:          
-                matchCategory = 'POSSIBLY_RELATED'
+            #--status display
+            entityCount += 1
+            batchEntityCount += 1
+            if entityCount % progressInterval == 0 or entity_row[0] == last_row_entity_id:
+                threadsRunning = 0
+                for process in process_list:
+                    if process.is_alive():
+                        threadsRunning += 1
+                qdepth = entity_queue.qsize() + resume_queue.qsize()
+                now = datetime.now().strftime('%I:%M%p').lower()
+                elapsedMins = round((time.time() - procStartTime) / 60, 1)
+                eps = int(float(entityCount) / (float(time.time() - procStartTime if time.time() - procStartTime != 0 else 1)))
+                eps2 = int(float(batchEntityCount) / (float(time.time() - batchStartTime if time.time() - batchStartTime != 0 else 1)))
+                print('%s entities processed at %s after %s minutes, %s / %s per second, %s processes, %s entity queue, %s resume queue' % (entityCount, now, elapsedMins, eps2, eps, threadsRunning, entity_queue.qsize(), resume_queue.qsize()))
+                batchStartTime = time.time()
+                batchEntityCount = 0
 
-            if relatedID not in resumeData:
-                resumeData[relatedID] = {}
-                resumeData[relatedID]['matchCategory'] = matchCategory
-                resumeData[relatedID]['dataSources'] = {}
+            #--get out if errors hit or out of records
+            if shutDown.value:
+                if shutDown.value == 9:
+                    print('USER INTERUPT! Shutting down ... ')
+                break
+ 
+        #--get out if errors hit
+        if shutDown.value:
+            break
+        else: #--set next batch
 
-            if dataSource not in resumeData[relatedID]['dataSources']:
-                resumeData[relatedID]['dataSources'][dataSource] = 1
-            else:
-                resumeData[relatedID]['dataSources'][dataSource] += 1
-
-            if exportFilePath and recordID: #--filters json related entities which currently only provide a summart, not a list of records
-                writeCsvRecord(rowData)
-
-            if shutDown:
+            if endEntityId >= maxEntityId:
                 break
 
-        #--update entity size breakdown
-        randomIndex = randomSampleIndex()
-        strEntitySize = str(entitySize)
-        if strEntitySize not in statPack['ENTITY_SIZE_BREAKDOWN']:
-            statPack['ENTITY_SIZE_BREAKDOWN'][strEntitySize] = {}
-            statPack['ENTITY_SIZE_BREAKDOWN'][strEntitySize]['COUNT'] = 0
-            statPack['ENTITY_SIZE_BREAKDOWN'][strEntitySize]['SAMPLE'] = []
-        statPack['ENTITY_SIZE_BREAKDOWN'][strEntitySize]['COUNT'] += 1
-        if len(statPack['ENTITY_SIZE_BREAKDOWN'][strEntitySize]['SAMPLE']) < sampleSize:
-            statPack['ENTITY_SIZE_BREAKDOWN'][strEntitySize]['SAMPLE'].append(entityID)
-            entitySizeReviewCount += 1
-        elif randomIndex != 0:
-            statPack['ENTITY_SIZE_BREAKDOWN'][strEntitySize]['SAMPLE'][randomIndex] = entityID
+            #--write interim snapshot file
+            queuesEmpty = wait_for_queues(entity_queue, resume_queue)
+            statData = {
+                'writeStatus': 'Interim',
+                'lastEntityId': endEntityId,
+                'queuesEmpty': queuesEmpty,
+                'statsFileName': statsFilePath
+            }
+            queue_write(resume_queue, statData)
+            queuesEmpty = wait_for_queues(entity_queue, resume_queue)
 
-        #--resolved entity stats
-        statPack['TOTAL_ENTITY_COUNT'] += 1
-        statPack['TOTAL_RECORD_COUNT'] += entitySize
-        for dataSource1 in resumeData['0']['dataSources']:
-            recordCount = resumeData['0']['dataSources'][dataSource1]
+            #--get next chunk
+            begEntityId += chunkSize
+            endEntityId += chunkSize
 
-            #--this just updates entity and record count for the data source
-            updateStatpack(dataSource1, None, None, 1, recordCount, None) 
+    #--write final snapshot file
+    print('Finishing up ...')
+    queuesEmpty = wait_for_queues(entity_queue, resume_queue)
+    if not shutDown.value:
+        statData = {
+            'writeStatus': 'Final',
+            'lastEntityId': maxEntityId if not datasourceFilter else datasourceFilter,
+            'queuesEmpty': queuesEmpty,
+            'statsFileName': statsFilePath
+        }
+        queue_write(resume_queue, statData)
+        queuesEmpty = wait_for_queues(entity_queue, resume_queue)
 
-            if recordCount == 1:
-                updateStatpack(dataSource1, None, 'SINGLE', 1, 0, entityID)
-            else:
-                updateStatpack(dataSource1, None, 'DUPLICATE', 1, recordCount, entityID)
-
-            #--cross matches
-            for dataSource2 in resumeData['0']['dataSources']:
-                if dataSource2 == dataSource1:
-                    continue
-                updateStatpack(dataSource1, dataSource2, 'MATCH', 1, recordCount, entityID)
-
-            #--related entity stats
-            for relatedID in resumeData:
-                if relatedID == '0':
-                    continue
-                matchCategory = resumeData[relatedID]['matchCategory']
-                statPack[categoryTotalStat[matchCategory]] += 1
-                for dataSource2 in resumeData[relatedID]['dataSources']:
-                    recordCount = resumeData[relatedID]['dataSources'][dataSource2]
-                    if dataSource2 == dataSource1:
-                        dataSource2 = None
-
-                    #--avoid double counting within data source (can't be avoided across data sources)
-                    if entityID < relatedID or dataSource2:
-                        updateStatpack(dataSource1, dataSource2, matchCategory, 1, recordCount, entityID + ' ' + relatedID)
-
-        #--status display
-        batchEntitySizeSum += entitySize
-        if entitySize > maxEntitySize:
-            maxEntitySize = entitySize 
-        relationCount = len(resumeData) - 1
-        batchRelationCountSum += relationCount
-        if relationCount > maxRelationCount:
-            maxRelationCount = relationCount
-
-        if statPack['TOTAL_ENTITY_COUNT'] % progressInterval == 0 or not exportRecord:
-            entityCount = statPack['TOTAL_ENTITY_COUNT']
-            now = datetime.now().strftime('%I:%M%p').lower()
-            elapsedMins = round((time.time() - procStartTime) / 60, 1)
-            eps = int(float(progressInterval) / (float(time.time() - batchStartTime if time.time() - batchStartTime != 0 else 1)))
-            avgEntitySize = round(batchEntitySizeSum / progressInterval)
-            avgRelationCount = round(batchRelationCountSum / progressInterval)
-
-            if exportRecord:
-                print(' %s entities processed at %s, %s per second | avg/max entity size = %s/%s | avg/max relationship count %s/%s' % (entityCount, now, eps, avgEntitySize, maxEntitySize, avgRelationCount, maxRelationCount))
-            else:
-                eps = int(float(entityCount) / (float(time.time() - procStartTime if time.time() - procStartTime != 0 else 1)))
-                print(' %s entities completed at %s after %s minutes, at %s per second' % (entityCount, now, elapsedMins, eps))
-
-            batchStartTime = time.time()
-            batchEntitySizeSum = 0
-            batchRelationCountSum = 0
-            maxEntitySize = 0
-            maxRelationCount = 0
-
-        #--get out if errors hit or out of records
-        if shutDown or not rowData:
+    #--stop the threads
+    print('stopping threads ...')
+    with threadStop.get_lock():
+        threadStop.value = 1
+    start = time.time()
+    while time.time() - start <= 15:
+        if not any(process.is_alive() for process in process_list):
             break
- 
-    #--get out if errors hit
-    if shutDown:
-        return 1
+        time.sleep(1)
+    for process in process_list:
+        if process.is_alive():
+            print(process.name, 'did not terminate gracefully')
+            process.terminate()
+        process.join()
+    entity_queue.close() 
+    resume_queue.close() 
+
+#---------------------------------------
+def wait_for_queues(entity_queue, resume_queue):
+    waits = 0
+    while entity_queue.qsize() or resume_queue.qsize():
+        time.sleep(1)
+        waits += 1
+        if waits >= 10:
+            break
+        elif entity_queue.qsize() or resume_queue.qsize():
+            print(' waiting for %s entity_queue and %s resume_queue records' % (entity_queue.qsize(), resume_queue.qsize()))
+
+    if (entity_queue.qsize() or resume_queue.qsize()):
+        print(' warning: queues are not empty!')
+        return False
+    return True
+
+#---------------------------------------
+def write_stat_pack(statPack, statData):
+    writeStatus = statData['writeStatus']
+    lastEntityId = statData['lastEntityId']
+    queuesEmpty = statData['queuesEmpty']
+    statsFileName = statData['statsFileName']
+
+    print(' %s stats written to %s' % (writeStatus, statsFileName))
+
+    if writeStatus == 'Interim':
+        statPack['PROCESS']['STATUS'] = 'Interim'
+    elif shutDown.value == 0:
+        statPack['PROCESS']['STATUS'] = 'Complete'
+    elif shutDown.value == 9:
+        statPack['PROCESS']['STATUS'] = 'Aborted by user'
+    else:
+        statPack['PROCESS']['STATUS'] = 'Error!'
+
+    if lastEntityId != 0:
+        statPack['PROCESS']['LAST_ENTITY_ID'] = lastEntityId
+    statPack['PROCESS']['QUEUES'] = 'empty' if queuesEmpty else 'NOT EMPTY!'
+    statPack['PROCESS']['END_TIME'] = datetime.now().strftime('%m/%d/%Y %H:%M:%S')
+
+    diff = datetime.now() - datetime.strptime(statPack['PROCESS']['START_TIME'], '%m/%d/%Y %H:%M:%S')
+    days, seconds = diff.days, diff.seconds
+    hours = days * 24 + seconds // 3600
+    minutes = (seconds % 3600) // 60
+    seconds = seconds % 60
+    statPack['PROCESS']['RUN_TIME'] = '%s:%s:%s' % (hours, minutes, seconds)
+    #statPack['PROCESS']['ENTITIES_PER_SECOND'] = int(float(statPack['TOTAL_ENTITY_COUNT']) / float(diff.seconds)) 
+    #--actual final stat
+    #elapsedMins = diff = data2 - data1
+    #eps = int(float(entityCount) / (float(time.time() - procStartTime if time.time() - procStartTime != 0 else 1)))
 
     #--calculate some percentages
-    statPack['TOTAL_COMPRESSION'] = str(round(100.00-((float(statPack['TOTAL_ENTITY_COUNT']) / float(statPack['TOTAL_RECORD_COUNT'])) * 100.00), 2)) + '%'
+    if statPack['TOTAL_RECORD_COUNT']:
+        statPack['TOTAL_COMPRESSION'] = str(round(100.00-((float(statPack['TOTAL_ENTITY_COUNT']) / float(statPack['TOTAL_RECORD_COUNT'])) * 100.00), 2)) + '%'
     for dataSource in statPack['DATA_SOURCES']:
         statPack['DATA_SOURCES'][dataSource]['COMPRESSION'] = str(round(100.00-((float(statPack['DATA_SOURCES'][dataSource]['ENTITY_COUNT']) / float(statPack['DATA_SOURCES'][dataSource]['RECORD_COUNT'])) * 100.00), 2)) + '%'
 
-    #--add feature stats to the entity size break down
-    print('\nReviewing %s entities ...' % entitySizeReviewCount)
-    reviewStartTime = time.time()
-    batchStartTime = time.time()
-    reviewCount = 0
-    getFlags = g2Engine.G2_ENTITY_INCLUDE_REPRESENTATIVE_FEATURES
-    entitySizeBreakdown = {}
-    for strEntitySize in statPack['ENTITY_SIZE_BREAKDOWN']:
-        entitySize = int(strEntitySize)
-        if entitySize < 10:
-            entitySizeLevel = entitySize
-        elif entitySize < 100:
-            entitySizeLevel = int(entitySize/10) * 10
-        else:
-            entitySizeLevel = int(entitySize/100) * 100
+    with open(statsFileName, 'w') as outfile:
+        json.dump(statPack, outfile, indent=4)    
 
-        if entitySizeLevel not in entitySizeBreakdown:
-            entitySizeBreakdown[entitySizeLevel] = {}
-            entitySizeBreakdown[entitySizeLevel]['ENTITY_COUNT'] = 0
-            entitySizeBreakdown[entitySizeLevel]['SAMPLE_ENTITIES'] = []
-            entitySizeBreakdown[entitySizeLevel]['REVIEW_COUNT'] = 0
-            entitySizeBreakdown[entitySizeLevel]['REVIEW_FEATURES'] = []
-        entitySizeBreakdown[entitySizeLevel]['ENTITY_COUNT'] += statPack['ENTITY_SIZE_BREAKDOWN'][strEntitySize]['COUNT']
-
-        for entityID in statPack['ENTITY_SIZE_BREAKDOWN'][strEntitySize]['SAMPLE']:
-
-            #--gather feature statistics
-            entityInfo = {}
-            if entitySize > 1:
-                try: 
-                    response = bytearray()
-                    retcode = g2Engine.getEntityByEntityIDV2(int(entityID), getFlags, response)
-                    response = response.decode() if response else ''
-                except G2Exception as err:
-                    print(str(err))
-                    #shutDown = True
-                    #return 1
-                    continue
-                try: 
-                    jsonData = json.loads(response)
-                except:
-                    response = 'None' if not response else response
-                    print('warning: entity %s response %s' % (entityID, response))
-                    continue
-
-                for ftypeCode in jsonData['RESOLVED_ENTITY']['FEATURES']:
-
-                    #-count how many with special cases 
-                    distinctFeatureCount = 0
-                    for distinctFeature in jsonData['RESOLVED_ENTITY']['FEATURES'][ftypeCode]:
-                        if ftypeCode == 'GENDER' and distinctFeature['FEAT_DESC'] not in ('M', 'F'): #--don't count invalid genders
-                            continue
-                        distinctFeatureCount += 1
-
-                    if ftypeCode not in entityInfo:
-                        entityInfo[ftypeCode] = 0
-                    entityInfo[ftypeCode] += distinctFeatureCount
-
-                if entitySize <= 3: #--super small
-                    maxExclusiveCnt = 1
-                    maxNameCnt = 2
-                    maxAddrCnt = 2
-                    #maxF1Cnt = 3
-                    #maxFFCnt = 5
-                elif entitySize <= 10: #--small
-                    maxExclusiveCnt = 1
-                    maxNameCnt = 3
-                    maxAddrCnt = 3
-                    #maxF1Cnt = 3
-                    #maxFFCnt = 5
-                elif entitySize <= 50: #--medium
-                    maxExclusiveCnt = 1
-                    maxNameCnt = 10
-                    maxAddrCnt = 10
-                    #maxF1Cnt = 10
-                    #maxFFCnt = 10
-                else: #--large
-                    maxExclusiveCnt = 1 #--large
-                    maxNameCnt = 25
-                    maxAddrCnt = 25
-                    #maxF1Cnt = 25
-                    #maxFFCnt = 25
-
-                reviewFeatures = []
-                for ftypeCode in entityInfo:
-                    distinctFeatureCount = entityInfo[ftypeCode]
-
-                    #--watch lists have more multiple features per record like 5 dobs and 10 names!
-                    if distinctFeatureCount > entitySize:
-                        continue
-
-                    frequency = ftypeCodeLookup[ftypeCode]['FTYPE_FREQ']
-                    exclusive = str(ftypeCodeLookup[ftypeCode]['FTYPE_EXCL']).upper() in ('1', 'Y', 'YES')
-
-                    needsReview = False
-                    if exclusive and distinctFeatureCount > maxExclusiveCnt:
-                        needsReview = True
-                    elif ftypeCode == 'NAME' and distinctFeatureCount > maxNameCnt:
-                        needsReview = True
-                    elif ftypeCode == 'ADDRESS' and distinctFeatureCount > maxAddrCnt:
-                        needsReview = True
-
-                    if needsReview: 
-                        reviewFeatures.append(ftypeCode)
-
-                if reviewFeatures:
-                    entityInfo['REVIEW_FEATURES'] = reviewFeatures
-                    entitySizeBreakdown[entitySizeLevel]['REVIEW_COUNT'] += 1
-                    for ftypeCode in reviewFeatures:
-                        if ftypeCode not in entitySizeBreakdown[entitySizeLevel]['REVIEW_FEATURES']:
-                            entitySizeBreakdown[entitySizeLevel]['REVIEW_FEATURES'].append(ftypeCode)
-
-            entityInfo['ENTITY_ID'] = entityID
-            entityInfo['ENTITY_SIZE'] = entitySize
-            entitySizeBreakdown[entitySizeLevel]['SAMPLE_ENTITIES'].append(entityInfo)
-
-            #--status display
-            reviewCount += 1
-            if reviewCount % progressInterval == 0 or reviewCount == entitySizeReviewCount:
-                entityCount = reviewCount
-                now = datetime.now().strftime('%I:%M%p').lower()
-                elapsedMins = round((time.time() - procStartTime) / 60, 1)
-                eps = int(float(progressInterval) / (float(time.time() - batchStartTime if time.time() - batchStartTime != 0 else 1)))
-                batchStartTime = time.time()
-                if rowData:
-                    print(' %s entities reviewed at %s, %s per second' % (entityCount, now, eps))
-                else:
-                    eps = int(float(entitySizeReviewCount) / (float(time.time() - reviewStartTime if time.time() - reviewStartTime != 0 else 1)))
-                    print(' %s entities reviewed at %s after %s minutes' % (entityCount, now, elapsedMins))
-
-    statPack['ENTITY_SIZE_BREAKDOWN'] = []
-    for entitySize in sorted(entitySizeBreakdown.keys()):
-        entitySizeRecord = entitySizeBreakdown[entitySize]
-        entitySizeRecord['ENTITY_SIZE'] = int(entitySize)
-        entitySizeRecord['ENTITY_SIZE_GROUP'] = str(entitySize) + ('+' if int(entitySize) >= 10 else '')
-        statPack['ENTITY_SIZE_BREAKDOWN'].append(entitySizeRecord)
-
-    return 0
+    return statPack
 
 #----------------------------------------
 def signal_handler(signal, frame):
-    print('USER INTERUPT! Shutting down ... (please wait)')
-    global shutDown
-    shutDown = True
+    with shutDown.get_lock():
+        shutDown.value = 9
     return
 
 #----------------------------------------
@@ -526,55 +793,86 @@ def pause(question='PRESS ENTER TO CONTINUE ...'):
 if __name__ == '__main__':
     appPath = os.path.dirname(os.path.abspath(sys.argv[0]))
 
-    global shutDown
-    shutDown = False
+    shutDown = Value('i', 0)
+    threadStop = Value('i', 0)
     signal.signal(signal.SIGINT, signal_handler)
+
     procStartTime = time.time()
     progressInterval = 10000
 
     #--defaults
-    try: iniFileName = G2Paths.get_G2Module_ini_path()
-    except: iniFileName = '' 
+    try: configFileName = G2Paths.get_G2Module_ini_path()
+    except: configFileName = '' 
     outputFileRoot = os.getenv('SENZING_OUTPUT_FILE_ROOT') if os.getenv('SENZING_OUTPUT_FILE_ROOT', None) else None
     sampleSize = int(os.getenv('SENZING_SAMPLE_SIZE')) if os.getenv('SENZING_SAMPLE_SIZE', None) and os.getenv('SENZING_SAMPLE_SIZE').isdigit() else 1000
+    datasourceFilter = os.getenv('SENZING_DATASOURCE_FILTER', None)
     relationshipFilter = int(os.getenv('SENZING_RELATIONSHIP_FILTER')) if os.getenv('SENZING_RELATIONSHIP_FILTER', None) and os.getenv('SENZING_RELATIONSHIP_FILTER').isdigit() else 3
+    chunkSize = int(os.getenv('SENZING_CHUNK_SIZE')) if os.getenv('SENZING_CHUNK_SIZE', None) and os.getenv('SENZING_CHUNK_SIZE').isdigit() else 1000000
+    threadCount = int(os.getenv('SENZING_THREAD_COUNT')) if os.getenv('SENZING_THREAD_COUNT', None) and os.getenv('SENZING_THREAD_COUNT').isdigit() else 0
 
     #--capture the command line arguments
     argParser = argparse.ArgumentParser()
-    argParser.add_argument('-c', '--config_file_name', dest='ini_file_name', default=iniFileName, help='name of the g2.ini file, defaults to %s' % iniFileName)
-    argParser.add_argument('-o', '--output_file_root', dest='output_file_root', default=outputFileRoot, help='root name for files created such as "/project/snapshots/snapshot1"')
-    argParser.add_argument('-s', '--sample_size', dest='sample_size', type=int, default=sampleSize, help='defaults to %s' % sampleSize)
-    argParser.add_argument('-f', '--relationship_filter', dest='relationship_filter', type=int, default=relationshipFilter, help='filter options 1=No Relationships, 2=Include possible matches, 3=Include possibly related and disclosed. Defaults to %s' % relationshipFilter)
-    argParser.add_argument('-a', '--for_audit', dest='export_csv', action='store_true', default=False, help='export csv file for audit')
-    argParser.add_argument('-D', '--debug', dest='debug', action='store_true', default=False, help='print debug statements')
+    argParser.add_argument('-c', '--config_file_name', default=configFileName, help='name of the senzing config file, defaults to %s' % configFileName)
+    argParser.add_argument('-o', '--output_file_root', default=outputFileRoot, help='root name for files created such as "/project/snapshots/snapshot1"')
+    argParser.add_argument('-s', '--sample_size', type=int, default=sampleSize, help='defaults to %s' % sampleSize)
+    argParser.add_argument('-d', '--datasource_filter', help='data source code to analayze, defaults to all')
+    argParser.add_argument('-f', '--relationship_filter', type=int, default=relationshipFilter, help='filter options 1=No Relationships, 2=Include possible matches, 3=Include possibly related and disclosed. Defaults to %s' % relationshipFilter)
+    argParser.add_argument('-a', '--for_audit', action='store_true', default=False, help='export csv file for audit')
+    argParser.add_argument('-D', '--debug', action='store_true', default=False, help='print debug statements')
+    argParser.add_argument('-k', '--chunk_size', type=int, default=chunkSize, help='defaults to %s' % chunkSize)
+    argParser.add_argument('-t', '--thread_count', type=int, default=threadCount, help='defaults to %s' % threadCount)
+    argParser.add_argument('-u', '--use_api', action='store_true', default=False, help='use api instead of sql to get resume')
 
     args = argParser.parse_args()
-    iniFileName = args.ini_file_name
+    configFileName = args.config_file_name
     outputFileRoot = args.output_file_root
     sampleSize = args.sample_size
+    datasourceFilter = args.datasource_filter
     relationshipFilter = args.relationship_filter
-    noCsvExport = not args.export_csv
+    exportCsv = args.for_audit
     debugOn = args.debug
-    exportType = 'CSV'
+    chunkSize = args.chunk_size
+    threadCount = args.thread_count
+    use_api = args.use_api
 
-    #--try to initialize the g2engine
+    #--get parameters from ini file
+    if not os.path.exists(configFileName):
+        print('')
+        print('An ini file was not found, please supply with the -c parameter.')
+        print('')
+        sys.exit(1)
+    iniParser = configparser.ConfigParser()
+    iniParser.read(configFileName)
+    try: g2dbUri = iniParser.get('SQL', 'CONNECTION')
+    except: 
+        print('')
+        print('CONNECTION parameter not found in [SQL] section of the ini file')
+        print('')
+        sys.exit(1)
+
+    #--g2 engine stuff
     try:
-        g2Engine = G2Engine()
-        iniParamCreator = G2IniParams()
-        iniParams = iniParamCreator.getJsonINIParams(iniFileName)
-        g2Engine.initV2('G2Snapshot', iniParams, False)
-    except G2Exception as err:
+        g2iniParams = G2IniParams()
+        iniParams = g2iniParams.getJsonINIParams(configFileName)
+        g2Product = G2Product()
+        apiVersion = json.loads(g2Product.version())
+        g2Product.destroy()
+    except G2Exception.G2Exception as err:
         print('\n%s\n' % str(err))
         sys.exit(1)
 
-    #--get the version information
-    try: 
-        g2Product = G2Product()
-        apiVersion = json.loads(g2Product.version())
-    except G2Exception.G2Exception as err:
-        print(err)
-        sys.exit(1)
-    g2Product.destroy()
+    if not threadCount:
+        try:
+            g2Diag = G2Diagnostic()
+            g2Diag.initV2('pyG2Diagnostic', iniParams, False)
+            physical_cores = g2Diag.getPhysicalCores()
+            logical_cores = g2Diag.getLogicalCores()
+            calc_cores_factor = 4 #--if physical_cores != logical_cores else 1.2
+            threadCount = math.ceil(logical_cores * calc_cores_factor)
+            print(f'\nPhysical cores: {logical_cores}, logical cores: {logical_cores}, default threads: {threadCount}')
+        except G2Exception.G2Exception as err:
+            print('\n%s\n' % str(err))
+            sys.exit(1)
 
     #--get needed config data
     try: 
@@ -586,71 +884,70 @@ if __name__ == '__main__':
         g2ConfigMgr.getConfig(defaultConfigID, defaultConfigDoc)
         cfgData = json.loads(defaultConfigDoc.decode())
         g2ConfigMgr.destroy()
+
         ftypeCodeLookup = {}
         for cfgRecord in cfgData['G2_CONFIG']['CFG_FTYPE']:
             ftypeCodeLookup[cfgRecord['FTYPE_CODE']] = cfgRecord 
-    except Exception as err:
+
+        dsrcLookup = {}
+        dsrcLookupByCode = {}
+        for cfgRecord in cfgData['G2_CONFIG']['CFG_DSRC']:
+            dsrcLookup[cfgRecord['DSRC_ID']] = cfgRecord 
+            dsrcLookupByCode[cfgRecord['DSRC_CODE']] = cfgRecord
+
+        erruleLookup = {}
+        for cfgRecord in cfgData['G2_CONFIG']['CFG_ERRULE']:
+            erruleLookup[cfgRecord['ERRULE_ID']] = cfgRecord 
+
+        ftypeLookup = {}
+        ambiguousFtypeID = 0
+        for cfgRecord in cfgData['G2_CONFIG']['CFG_FTYPE']:
+            ftypeLookup[cfgRecord['FTYPE_ID']] = cfgRecord 
+            if cfgRecord['FTYPE_CODE'] == 'AMBIGUOUS_ENTITY':
+                ambiguousFtypeID = cfgRecord['FTYPE_ID']
+
+    except G2Exception.G2Exception as err:
         print('\n%s\n' % str(err))
         sys.exit(1)
     g2ConfigMgr.destroy()
 
+    #--validate data source filter if supplied
+    if datasourceFilter:
+        if datasourceFilter.upper() not in dsrcLookupByCode:
+            print(f'\nData source code {datasourceFilter} is not valid\n')
+            sys.exit(1)
+        else:
+            datasourceFilterID = dsrcLookupByCode[datasourceFilter.upper()]['DSRC_ID']
+
     #--check the output file
     if not outputFileRoot:
-        print('\nPlease use -o to select and output path and root file name such as /project/audit/run1\n')
+        print('\nPlease use -o to select and output path and root file name such as /project/audit/snap1\n')
         sys.exit(1)
     if os.path.splitext(outputFileRoot)[1]:
         print("\nPlease don't use a file extension as both a .json and a .csv file will be created\n")
         sys.exit(1)
 
-    #--create output file paths
     statsFilePath = outputFileRoot + '.json'
-    if noCsvExport:
-        exportFilePath = None
-    else:
-        exportFilePath = outputFileRoot + '.csv'
+    csvFilePath = outputFileRoot + '.csv'
 
-    #--open the export file if set
-    if exportFilePath:
-        columnHeaders = []
-        columnHeaders.append('RESOLVED_ENTITY_ID')
-        columnHeaders.append('RELATED_ENTITY_ID')
-        columnHeaders.append('MATCH_LEVEL')
-        columnHeaders.append('MATCH_KEY')
-        columnHeaders.append('DATA_SOURCE')
-        columnHeaders.append('RECORD_ID')
-        try: 
-            exportFileHandle = open(exportFilePath, 'w')
-            exportFileHandle.write(','.join(columnHeaders) + '\n')        
-        except IOError as err: 
-            print('\nERROR: cannot write to %s \n%s\n' % (exportFilePath, err))
-            sys.exit(1)
-            
-    #--get entities and relationships
-    statPack = {}
-    statPack['SOURCE'] = 'G2Snapshot'
-    statPack['API_VERSION'] = apiVersion['BUILD_VERSION']
-    statPack['RUN_DATE'] = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-    returnCode = processEntities()
+    statsFileExisted = os.path.exists(statsFilePath)
+    try:
+        with open(statsFilePath, 'a') as f:
+            pass
+    except IOError as err:
+        print(f'\nCannot write to {statsFilePath}.  \n{err}\n')
+        sys.exit(1)
+    if not statsFileExisted:
+        os.remove(statsFilePath)
 
-    #--wrap ups
-    if exportFilePath:
-        exportFileHandle.close()
-    g2Engine.destroy()
-
-    #--dump the stats to screen and file
-    print('')
-    for stat in statPack:
-        if type(statPack[stat]) not in (list, dict):
-            print ('%s = %s' % (stat, statPack[stat]))
-    with open(statsFilePath, 'w') as outfile:
-        json.dump(statPack, outfile)    
-    print('')
+    #--process the entities
+    processEntities()
 
     elapsedMins = round((time.time() - procStartTime) / 60, 1)
-    if returnCode == 0:
+    if shutDown.value == 0:
         print('Process completed successfully in %s minutes' % elapsedMins)
     else:
         print('Process aborted after %s minutes!' % elapsedMins)
     print('')
 
-    sys.exit(returnCode)
+    sys.exit(shutDown.value)
